Fix warnings
Ignore C26812

+++ src/program/OutfitStudio.cpp
@@ -5761,7 +5761,7 @@ void OutfitStudioFrame::ApplyPartitions() {
 		PartitionItemData* partitionData = dynamic_cast<PartitionItemData*>(partitionTree->GetItemData(child));
 		if (partitionData) {
 			const int index = partitionData->index;
-			if (index >= partitionInfo.size()) {
+			if (index >= static_cast<int64_t>(partitionInfo.size())) {
 				partitionInfo.resize(index + 1);
 				delPartFlags.resize(index + 1, true);
 			}
@@ -5807,8 +5807,8 @@ void OutfitStudioFrame::CreatePartitionTree(NiShape* shape) {
 
 	NiVector<BSDismemberSkinInstance::PartitionInfo> partitionInfo;
 	if (project->GetWorkNif()->GetShapePartitions(shape, partitionInfo, triParts)) {
-		for (int i = 0; i < partitionInfo.size(); i++) {
-			partitionTree->AppendItem(partitionRoot, "Partition", -1, -1, new PartitionItemData(i, partitionInfo[i].partID));
+		for (uint32_t i = 0; i < partitionInfo.size(); i++) {
+			partitionTree->AppendItem(partitionRoot, "Partition", -1, -1, new PartitionItemData(static_cast<int>(i), partitionInfo[i].partID));
 		}
 	}
